move projects sql builder to util, add streaming csv download endpoint
