+++ routes/projects.js
@@ -3,12 +3,16 @@ const path = require('path');
 const SphericalMercator = require('sphericalmercator');
 const NodeCache = require('node-cache');
 const shortid = require('shortid');
-const generateDynamicQuery = require('../utils/generate-dynamic-sql');
 const turfBuffer = require('@turf/buffer');
 const turfBbox = require('@turf/bbox');
 const { Recaptcha } = require('express-recaptcha');
 const github = require('octonode');
 const fetch = require('node-fetch');
+const Json2csvTransform = require('json2csv').Transform;
+const QueryStream = require('pg-query-stream');
+const JSONStream = require('JSONStream');
+
+const buildProjectsSQL = require('../utils/build-projects-sql');
 
 
 const recaptcha = new Recaptcha(process.env.RECAPTCHA_SITE_KEY, process.env.RECAPTCHA_SECRET_KEY);
@@ -42,112 +46,29 @@ function sql(file) {
 }
 
 // import sql query templates
-const listProjectsQuery = sql('../queries/projects/index.sql');
 const findProjectQuery = sql('../queries/projects/show.sql');
-const paginateQuery = sql('../queries/helpers/paginate.sql');
-const standardColumns = sql('../queries/helpers/standard-projects-columns.sql');
 const boundingBoxQuery = sql('../queries/helpers/bounding-box-query.sql');
 const generateVectorTile = sql('../queries/helpers/generate-vector-tile.sql');
 
-/* GET /projects */
 router.get('/', async (req, res) => {
-  // extract params, set defaults
-  const {
-    query: {
-      // pagination
-      page = '1',
-      itemsPerPage = 30,
+  const { query } = req;
 
-      // filters
-      'community-districts': communityDistricts = [],
-      'action-types': actionTypes = [],
-      boroughs = [],
-      dcp_ceqrtype = ['Type I', 'Type II', 'Unlisted', 'Unknown'],
-      dcp_ulurp_nonulurp = ['ULURP', 'Non-ULURP'],
-      dcp_femafloodzonev = false,
-      dcp_femafloodzonecoastala = false,
-      dcp_femafloodzonea = false,
-      dcp_femafloodzoneshadedx = false,
-      dcp_publicstatus = ['Complete', 'Filed', 'In Public Review', 'Unknown'],
-      dcp_certifiedreferred = [],
-      project_applicant_text = '',
-      ulurp_ceqr_text = '',
-      block = '',
-    },
-  } = req;
-
-
-  const paginate = generateDynamicQuery(paginateQuery, { itemsPerPage, offset: (page - 1) * itemsPerPage });
-  const communityDistrictsQuery =
-    communityDistricts[0] ? pgp.as.format('AND dcp_validatedcommunitydistricts ilike any (array[$1:csv])', [communityDistricts.map(district => `%${district}%`)]) : '';
-
-  const boroughsQuery = boroughs[0] ? pgp.as.format('AND dcp_borough ilike any (array[$1:csv])', [boroughs.map(borough => `%${borough}%`)]) : '';
-
-  const actionTypesQuery = actionTypes[0] ? pgp.as.format('AND actiontypes ilike any (array[$1:csv])', [actionTypes.map(actionType => `%${actionType}%`)]) : '';
-  const certDateQuery = (!!dcp_certifiedreferred[0] && !!dcp_certifiedreferred[1]) ? pgp.as.format('AND dcp_certifiedreferred BETWEEN to_timestamp($1) AND to_timestamp($2)', dcp_certifiedreferred) : '';
-  // special handling for FEMA flood zones
-  // to only filter when set to true
-  const dcp_femafloodzonevQuery = dcp_femafloodzonev === 'true' ? 'AND dcp_femafloodzonev = true' : '';
-  const dcp_femafloodzonecoastalaQuery = dcp_femafloodzonecoastala === 'true' ? 'AND dcp_femafloodzonecoastala = true' : '';
-  const dcp_femafloodzoneaQuery = dcp_femafloodzonea === 'true' ? 'AND dcp_femafloodzonea = true' : '';
-  const dcp_femafloodzoneshadedxQuery = dcp_femafloodzoneshadedx === 'true' ? 'AND dcp_femafloodzoneshadedx = true' : '';
-  const projectApplicantTextQuery = project_applicant_text ? pgp.as.format("AND ((dcp_projectbrief ilike '%$1:value%') OR (dcp_projectname ilike '%$1:value%') OR (dcp_applicant ilike '%$1:value%'))", [project_applicant_text]) : '';
-  const ulurpCeqrQuery = ulurp_ceqr_text ? pgp.as.format("AND ((ulurpnumbers ILIKE '%$1:value%') OR dcp_ceqrnumber ILIKE '%$1:value%')", [ulurp_ceqr_text]) : '';
-  const blockQuery = block ? pgp.as.format("AND (blocks ilike '%$1:value%')", [block]) : '';
+  const SQL = buildProjectsSQL(query);
 
   try {
-    const projects =
-      await db.any(listProjectsQuery, {
-        standardColumns,
-        dcp_publicstatus,
-        dcp_ceqrtype,
-        dcp_ulurp_nonulurp,
-        dcp_femafloodzonevQuery,
-        dcp_femafloodzonecoastalaQuery,
-        dcp_femafloodzoneaQuery,
-        dcp_femafloodzoneshadedxQuery,
-        certDateQuery,
-        communityDistrictsQuery,
-        boroughsQuery,
-        actionTypesQuery,
-        projectApplicantTextQuery,
-        ulurpCeqrQuery,
-        blockQuery,
-        paginate,
-      });
+    const projects = await db.any(SQL);
 
     const [{ total_projects: total = 0 } = {}] = projects || [];
     const { length = 0 } = projects;
 
     // if this is the first page of a new query, include bounds for the query's geoms, and a vector tile template
     let tileMeta = {};
-
+    const { page } = query;
     if (page === '1') {
-      // tileQuery is uses the same WHERE clauses as above,
-      // but only SELECTs geom, projectid, and projectname, and does not include pagination
-
-      const tileQuery = pgp.as.format(listProjectsQuery, {
-        standardColumns: 'geom, projectid, dcp_projectname, dcp_publicstatus_simp',
-        dcp_publicstatus,
-        dcp_ceqrtype,
-        dcp_ulurp_nonulurp,
-        dcp_femafloodzonevQuery,
-        dcp_femafloodzonecoastalaQuery,
-        dcp_femafloodzoneaQuery,
-        dcp_femafloodzoneshadedxQuery,
-        certDateQuery,
-        communityDistrictsQuery,
-        boroughsQuery,
-        actionTypesQuery,
-        projectApplicantTextQuery,
-        ulurpCeqrQuery,
-        blockQuery,
-        paginate: '',
-      });
+      const tileQuery = buildProjectsSQL(query, 'tiles');
 
       // create array of projects that have geometry
       const projectsWithGeometries = projects.filter(project => project.has_centroid);
-
 
       // get the bounds for projects with geometry
       // default to a bbox for the whole city
@@ -188,7 +109,6 @@ router.get('/', async (req, res) => {
       };
     }
 
-
     // send the response with a tile template
     res.send({
       data: projects.map(project => ({
@@ -208,6 +128,41 @@ router.get('/', async (req, res) => {
     });
   }
 });
+
+router.get('/download.csv', async (req, res) => {
+  const { query } = req;
+
+  const SQL = buildProjectsSQL(query, 'download');
+
+  // you can also use pgp.as.format(query, values, options)
+  // to format queries properly, via pg-promise;
+  const qs = new QueryStream(SQL);
+
+  const transformOpts = { highWaterMark: 16384, encoding: 'utf-8' };
+  const json2csv = new Json2csvTransform({}, transformOpts);
+
+  // Set approrpiate download headers
+  res.setHeader('Content-disposition', 'attachment; filename=projects.csv');
+  res.writeHead(200, { 'Content-Type': 'text/csv' });
+
+  // Flush the headers before we start pushing the CSV content
+  res.flushHeaders();
+
+  db.stream(qs, (s) => {
+    // initiate streaming into the console:
+    s.pipe(JSONStream.stringify()).pipe(json2csv).pipe(res);
+  })
+    .then((data) => {
+      console.log(
+        'Total rows processed:', data.processed,
+        'Duration in milliseconds:', data.duration,
+      );
+    })
+    .catch((error) => {
+      console.log('ERROR:', error);
+    });
+});
+
 
 /* GET /projects/:id */
 /* Retreive a single project */
